Restructure and add GIMP 3 upscaling plugin
Moved GIMP 2.10 upscaling plugin and RealESRGAN resources into a dedicated gimp2_upscale directory. Added a new gimp3_upscale directory with the GIMP 3 plugin and duplicated required RealESRGAN binaries and models. Removed the obsolete gimp_upscale_simple_2.10.py script.
